refactor: 리뷰 반영 및 폴더 구조 변경(#533)
